Fix pylint issue
